# swagger.yml
basePath: /
consumes:
    - application/json
definitions:
...
    ExtClient:
        description: ExtClient - struct for external clients
        properties:
            address:
                type: string
                x-go-name: Address
            address6:
                type: string
                x-go-name: Address6
            clientid:
                type: string
                x-go-name: ClientID
            deniednodeacls:
                additionalProperties:
                    type: object
                type: object
                x-go-name: DeniedACLs
            dns:
                type: string
                x-go-name: DNS
            enabled:
                type: boolean
                x-go-name: Enabled
            extraallowedips:
                items:
                    type: string
                type: array
                x-go-name: ExtraAllowedIPs
            ingressgatewayendpoint:
                type: string
                x-go-name: IngressGatewayEndpoint
            ingressgatewayid:
                type: string
                x-go-name: IngressGatewayID
            lastmodified:
                format: int64
                type: integer
                x-go-name: LastModified
            network:
                type: string
                x-go-name: Network
            ownerid:
                type: string
                x-go-name: OwnerID
            privatekey:
                type: string
                x-go-name: PrivateKey
            publickey:
                type: string
                x-go-name: PublicKey
            remote_access_client_id:
                type: string
                x-go-name: RemoteAccessClientID
        type: object
        x-go-package: github.com/gravitl/netmaker/models
...
host: api.demo.netmaker.io
info:
    description: |-
        # API Usage

        Most actions that can be performed via API can be performed via UI. We recommend managing your networks using the official netmaker-ui project. However, Netmaker can also be run without the UI, and all functions can be achieved via API calls. If your use case requires using Netmaker without the UI or you need to do some troubleshooting/advanced configuration, using the API directly may help.

        # Authentication

        API calls must be authenticated via a header of the format -H “Authorization: Bearer <YOUR_SECRET_KEY>” There are two methods to obtain YOUR_SECRET_KEY: 1. Using the masterkey. By default, this value is “secret key,” but you should change this on your instance and keep it secure. This value can be set via env var at startup or in a config file (config/environments/< env >.yaml). See the [Netmaker](https://docs.netmaker.org/index.html) documentation for more details. 2. Using a JWT received for a node. This can be retrieved by calling the /api/nodes/<network>/authenticate endpoint, as documented below.
    title: Netmaker
    version: 0.21.1
paths:
...
    /api/extclients:
        get:
            operationId: getAllExtClients
            parameters:
                - description: Networks
                  in: body
                  name: networks
                  schema:
                    items:
                        type: string
                    type: array
                  x-go-name: Networks
            responses:
                "200":
                    $ref: '#/responses/extClientSliceResponse'
            schemes:
                - https
            summary: A separate function to get all extclients, not just extclients for a particular network.
            tags:
                - ext_client
    /api/extclients/{network}:
        get:
            description: Gets all extclients associated with network, including pending extclients.
            operationId: getNetworkExtClients
            parameters:
                - description: Network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/extClientSliceResponse'
            schemes:
                - https
            summary: Get all extclients associated with network.
            tags:
                - ext_client
    /api/extclients/{network}/{clientid}:
        delete:
            operationId: deleteExtClient
            parameters:
                - description: Client ID
                  in: path
                  name: clientid
                  required: true
                  type: string
                  x-go-name: ClientID
                - description: Network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            schemes:
                - https
            summary: Delete an individual extclient.
            tags:
                - ext_client
        get:
            operationId: getExtClient
            parameters:
                - description: Client ID
                  in: path
                  name: clientid
                  required: true
                  type: string
                  x-go-name: ClientID
                - description: Network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/extClientResponse'
            schemes:
                - https
            summary: Get an individual extclient.
            tags:
                - ext_client
        put:
            operationId: updateExtClient
            parameters:
                - description: Client ID
                  in: path
                  name: clientid
                  required: true
                  type: string
                  x-go-name: ClientID
                - description: Network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - description: ExtClient
                  in: body
                  name: ext_client
                  schema:
                    $ref: '#/definitions/ExtClient'
                  x-go-name: ExtClient
            responses:
                "200":
                    $ref: '#/responses/extClientResponse'
            schemes:
                - https
            summary: Update an individual extclient.
            tags:
                - ext_client
    /api/extclients/{network}/{clientid}/{type}:
        get:
            operationId: getExtClientConf
            parameters:
                - description: Type
                  in: path
                  name: type
                  required: true
                  type: string
                  x-go-name: Type
                - description: Client ID
                  in: path
                  name: clientid
                  required: true
                  type: string
                  x-go-name: ClientID
                - description: Network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/extClientResponse'
            schemes:
                - https
            summary: Get an individual extclient.
            tags:
                - ext_client
    /api/extclients/{network}/{nodeid}:
        post:
            operationId: createExtClient
            parameters:
                - description: Network
                  in: path
                  name: network
                  required: true
                  type: string
                  x-go-name: Network
                - description: Node ID
                  in: path
                  name: nodeid
                  required: true
                  type: string
                  x-go-name: NodeID
                - description: Custom ExtClient
                  in: body
                  name: custom_ext_client
                  schema:
                    $ref: '#/definitions/CustomExtClient'
                  x-go-name: CustomExtClient
            responses:
                "200":
                    $ref: '#/responses/okResponse'
            schemes:
                - https
            summary: Create an individual extclient.  Must have valid key and be unique.
            tags:
                - ext_client
...
    /api/networks:
        get:
            operationId: getNetworks
            responses:
                "200":
                    $ref: '#/responses/getNetworksSliceResponse'
            schemes:
                - https
            summary: Lists all networks.
            tags:
                - networks
        post:
            operationId: createNetwork
            parameters:
                - description: Network
                  in: body
                  name: network
                  schema:
                    $ref: '#/definitions/Network'
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Create a network.
            tags:
                - networks
        put:
            operationId: updateNetwork
            parameters:
                - description: Network
                  in: body
                  name: network
                  schema:
                    $ref: '#/definitions/Network'
                  x-go-name: Network
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Update pro settings for a network.
            tags:
                - networks
    /api/networks/{networkname}:
        delete:
            operationId: deleteNetwork
            parameters:
                - description: 'name: network name'
                  in: path
                  name: networkname
                  required: true
                  type: string
                  x-go-name: Networkname
            responses:
                "200":
                    $ref: '#/responses/successResponse'
            schemes:
                - https
            summary: Delete a network.  Will not delete if there are any nodes that belong to the network.
            tags:
                - networks
        get:
            operationId: getNetwork
            parameters:
                - description: 'name: network name'
                  in: path
                  name: networkname
                  required: true
                  type: string
                  x-go-name: Networkname
            responses:
                "200":
                    $ref: '#/responses/networkBodyResponse'
            schemes:
                - https
            summary: Get a network.
            tags:
                - networks
...
    /api/users/{username}:
        delete:
            operationId: deleteUser
            parameters:
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Delete a user.
            tags:
                - user
        get:
            operationId: getUser
            parameters:
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Get an individual user.
            tags:
                - user
        post:
            operationId: createUser
            parameters:
                - description: User
                  in: body
                  name: user
                  schema:
                    $ref: '#/definitions/User'
                  x-go-name: User
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Create a user.
            tags:
                - user
        put:
            operationId: updateUser
            parameters:
                - description: User
                  in: body
                  name: user
                  schema:
                    $ref: '#/definitions/User'
                  x-go-name: User
                - description: Username
                  in: path
                  name: username
                  required: true
                  type: string
                  x-go-name: Username
            responses:
                "200":
                    $ref: '#/responses/userBodyResponse'
            schemes:
                - https
            summary: Update a user.
            tags:
                - user
...
produces:
    - application/json
responses:
...
    extClientResponse:
        description: ""
        schema:
            $ref: '#/definitions/ExtClient'
    extClientSliceResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/ExtClient'
            type: array
...
    getNetworksSliceResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/Network'
            type: array
...
    networkBodyResponse:
        description: ""
        schema:
            $ref: '#/definitions/Network'
...
    okResponse:
        description: ""
...
    successResponse:
        description: ""
        schema:
            $ref: '#/definitions/SuccessResponse'
...
schemes:
    - https
swagger: "2.0"